requirements.txt
Django==1.4
South==0.7.5
-e git://github.com/earle/django-bootstrap.git@2433b7df07f04a3a7c2ca5adf2c49f882a53ffa3#egg=bootstrap-dev
django-grappelli==2.3.8
django-tastypie==0.9.11
mimeparse==0.1.3
psycopg2==2.4.5
python-dateutil==1.5
wsgiref==0.1.2
